Split logo and runtime-version-selector into components
